+++ tikv-client-pd/Cargo.toml
@@ -5,13 +5,11 @@ edition = "2018"
 
 [dependencies]
 async-trait = "0.1"
-derive-new = "0.5"
 futures = { version = "0.3.5", features = ["compat", "async-await", "thread-pool"] }
 grpcio = { version = "0.6", features = [ "secure", "prost-codec" ], default-features = false }
 log = "0.4"
 tikv-client-common = { path = "../tikv-client-common" }
 tikv-client-proto = { path = "../tikv-client-proto" }
-tokio = { version = "0.2", features = ["sync"] }
 
 [dev-dependencies]
 clap = "2.32"
@@ -19,4 +17,3 @@ fail = { version = "0.3", features = [ "failpoints" ] }
 proptest = "0.9"
 proptest-derive = "0.1.0"
 tempdir = "0.3"
-tokio = { version = "0.2", features = ["rt-threaded", "macros"] }
